backend/refactor: decoupled Distance module
